Rename tests for receivers
