+++ kmock-processor/src/test/kotlin/tech/antibytes/kmock/integration/KMockMocksSpec.kt
@@ -945,13 +945,13 @@ class KMockMocksSpec {
     }
 
     @Test
-    fun `Given a annotated Source for a Platform which contains scoped Extensions is processed, it writes a mock`() {
+    fun `Given a annotated Source for a Platform which contains Receivers is processed, it writes a mock`() {
         // Given
         val source = SourceFile.kotlin(
             "Platform.kt",
-            loadResource("/template/scoped/Platform.kt")
+            loadResource("/template/receiver/Platform.kt")
         )
-        val expected = loadResource("/expected/scoped/Platform.kt")
+        val expected = loadResource("/expected/receiver/Platform.kt")
 
         // When
         val compilerResult = compile(
@@ -969,13 +969,13 @@ class KMockMocksSpec {
     }
 
     @Test
-    fun `Given a annotated Source for Shared which contains scoped Extensions is processed, it writes a mock`() {
+    fun `Given a annotated Source for Shared which contains Receivers is processed, it writes a mock`() {
         // Given
         val source = SourceFile.kotlin(
             "Shared.kt",
-            loadResource("/template/scoped/Shared.kt")
+            loadResource("/template/receiver/Shared.kt")
         )
-        val expected = loadResource("/expected/scoped/Shared.kt")
+        val expected = loadResource("/expected/receiver/Shared.kt")
 
         // When
         val compilerResult = compile(
@@ -993,13 +993,13 @@ class KMockMocksSpec {
     }
 
     @Test
-    fun `Given a annotated Source for Common which contains scoped Extensions is processed, it writes a mock`() {
+    fun `Given a annotated Source for Common which contains Receivers is processed, it writes a mock`() {
         // Given
         val source = SourceFile.kotlin(
             "Common.kt",
-            loadResource("/template/scoped/Common.kt")
+            loadResource("/template/receiver/Common.kt")
         )
-        val expected = loadResource("/expected/scoped/Common.kt")
+        val expected = loadResource("/expected/receiver/Common.kt")
 
         // When
         val compilerResult = compile(
