+++ kmock-processor/src/test/resources/mock/expected/receiver/Platform.kt
@@ -1,4 +1,4 @@
-package mock.template.scoped
+package mock.template.`receiver`
 
 import kotlin.Any
 import kotlin.Boolean
@@ -41,14 +41,14 @@ internal class PlatformMock(
         get() = _myThing.onGet()
 
     public val _myThing: KMockContract.PropertyProxy<String> =
-        ProxyFactory.createPropertyProxy("mock.template.scoped.PlatformMock#_myThing", collector =
+        ProxyFactory.createPropertyProxy("mock.template.receiver.PlatformMock#_myThing", collector =
         verifier, freeze = freeze)
 
     public override val AnythingElse.SomethingInside.inside: Int
         get() = throw IllegalStateException("This action is not callable.")
 
     public val _iDo: KMockContract.SyncFunProxy<Unit, () -> kotlin.Unit> =
-        ProxyFactory.createSyncFunProxy("mock.template.scoped.PlatformMock#_iDo", collector =
+        ProxyFactory.createSyncFunProxy("mock.template.receiver.PlatformMock#_iDo", collector =
         verifier, freeze = freeze)
 
     public override fun Something.equals(): Int = throw IllegalStateException(
